requirements.txt
psutil
pywin32
watchdog
python-telegram-bot==20.3
